Update example injection with namespace (#232)
The spring-petclinic example for setting up java auto instrumentation
using the operator didn't include kubernetes namespaces in the
injection commands. This would result in commands failing, due to not 
being able to find the deployment. This change adds namespace to
the commands so they'll be able to find the proper deployments.